% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanet_read_cv.R
\name{epanet_read_cv}
\alias{epanet_read_cv}
\title{Read Control Valves values and results}
\usage{
epanet_read_cv(hmodel, hreport, cv_name)
}
\arguments{
\item{hmodel}{object with entries for sections of the EPANET .inp file.}

\item{hreport}{object with entries for nodes and links of the EPANET .rpt file.}

\item{cv_name}{Name of the Control Valve}
}
\value{
The table operation_data
}
\description{
Read the operation data for the Control Valve
The data is extracted from the output report produced from a simulation.
}
